requirements.txt
langchain==0.3.1
langchain-core==0.3.9
langchain-community==0.3.1
langchain-openai==0.2.2
python-dotenv==1.0.1
openai==1.50.2
numexpr==2.10.1
httpx==0.27.0

